+++ isso/config.py
@@ -7,12 +7,7 @@ import logging
 import datetime
 
 from email.utils import parseaddr, formataddr
-try:
-    from backports.configparser import ConfigParser
-except ImportError:
-    from configparser import ConfigParser
-
-from isso.compat import text_type as str
+from configparser import ConfigParser, NoOptionError, NoSectionError
 
 logger = logging.getLogger("isso")
 
@@ -123,8 +118,9 @@ def new(options=None):
 def load(default, user=None):
 
     # return set of (section, option)
-    def setify(cp): return set((section, option) for section in cp.sections()
-                               for option in cp.options(section))
+    def setify(cp):
+        return set((section, option) for section in cp.sections()
+                   for option in cp.options(section))
 
     parser = new()
     parser.read(default)
@@ -144,8 +140,12 @@ def load(default, user=None):
             logger.info("Your `session-key` has been stored in the "
                         "database itself, this option is now unused")
 
-    if not parseaddr(parser.get("smtp", "from"))[0]:
+    try:
+        fromaddr = parser.get("smtp", "from")
+    except (NoOptionError, NoSectionError):
+        fromaddr = ''
+    if not parseaddr(fromaddr)[0]:
         parser.set("smtp", "from",
-                   formataddr(("Ich schrei sonst!", parser.get("smtp", "from"))))
+                   formataddr(("Ich schrei sonst!", fromaddr)))
 
     return parser