+++ isso/core.py
@@ -12,14 +12,10 @@ try:
 except ImportError:
     uwsgi = None
 
-from isso.compat import PY2K
+import _thread as thread
 
-if PY2K:
-    import thread
-else:
-    import _thread as thread
-
-from werkzeug.contrib.cache import NullCache, SimpleCache
+from flask_caching.backends.nullcache import NullCache
+from flask_caching.backends.simplecache import SimpleCache
 
 logger = logging.getLogger("isso")
 
@@ -123,7 +119,8 @@ class uWSGIMixin(Mixin):
 
         timedelta = conf.getint("moderation", "purge-after")
 
-        def purge(signum): return self.db.comments.purge(timedelta)
+        def purge(signum):
+            return self.db.comments.purge(timedelta)
         uwsgi.register_signal(1, "", purge)
         uwsgi.add_timer(1, timedelta)