+++ isso/utils/__init__.py
@@ -14,7 +14,6 @@ from jinja2 import Environment, FileSystemLoader
 from werkzeug.wrappers import Response
 from werkzeug.exceptions import BadRequest
 
-from isso.compat import text_type
 from isso.wsgi import Request
 
 try:
@@ -29,7 +28,7 @@ def anonymize(remote_addr):
     and /48 (zero'd).
 
     """
-    if not isinstance(remote_addr, text_type) and isinstance(remote_addr, str):
+    if not isinstance(remote_addr, str) and isinstance(remote_addr, str):
         remote_addr = remote_addr.decode('ascii', 'ignore')
     try:
         ipv4 = ipaddress.IPv4Address(remote_addr)
@@ -38,7 +37,7 @@ def anonymize(remote_addr):
         try:
             ipv6 = ipaddress.IPv6Address(remote_addr)
             if ipv6.ipv4_mapped is not None:
-                return anonymize(text_type(ipv6.ipv4_mapped))
+                return anonymize(str(ipv6.ipv4_mapped))
             return u'' + ipv6.exploded.rsplit(':', 5)[0] + ':' + ':'.join(['0000'] * 5)
         except ipaddress.AddressValueError:
             return u'0.0.0.0'