+++ isso/views/__init__.py
@@ -12,7 +12,6 @@ from werkzeug.routing import Rule
 from werkzeug.exceptions import BadRequest
 
 from isso import local
-from isso.compat import text_type as str
 
 
 class requires:
@@ -64,3 +63,14 @@ class Info(object):
         }
 
         return Response(json.dumps(rv), 200, content_type="application/json")
+
+
+class Metrics(object):
+
+    def __init__(self, isso):
+        isso.urls.add(Rule('/metrics', endpoint=self.show))
+
+    def show(self, environ, request):
+        content_type = 'text/plain; version=0.0.4; charset=utf-8'
+        metrics = ""
+        return Response(metrics, 200, content_type=content_type)